+++ lua/terrortown/entities/items/item_ttt_disguiser.lua
@@ -9,7 +9,6 @@ if SERVER then
 	AddCSLuaFile()
 end
 
-ITEM.hud = Material("vgui/ttt/perks/hud_disguiser.png")
 ITEM.EquipMenuData = {
 	type = "item_active",
 	name = "item_disg",
@@ -60,33 +59,6 @@ if CLIENT then
 		return dform
 	end
 
-	---
-	-- Draws the disguise info on screen for the disguised @{Player}
-	-- @param Player client the client
-	-- @realm client
-	function DISGUISE.Draw(client)
-		if not client or not client:IsActive() or not client:GetNWBool("disguised", false) then return end
-
-		trans = trans or LANG.GetTranslation
-
-		surface.SetFont("TabLarge")
-		surface.SetTextColor(255, 0, 0, 230)
-
-		local text = trans("disg_hud")
-		local _, h = surface.GetTextSize(text)
-
-		surface.SetTextPos(36, ScrH() - 160 - h)
-		surface.DrawText(text)
-	end
-
-	hook.Add("HUDPaint", "TTTItemDisguiser", function()
-		local client = LocalPlayer()
-
-		if client:Alive() and client:Team() ~= TEAM_SPEC and hook.Call("HUDShouldDraw", GAMEMODE, "TTTDisguise") then
-			DISGUISE.Draw(client)
-		end
-	end)
-
 	hook.Add("TTTEquipmentTabs", "TTTItemDisguiser", function(dsheet)
 		trans = trans or LANG.GetTranslation
 
@@ -96,7 +68,16 @@ if CLIENT then
 
 		dsheet:AddSheet(trans("disg_name"), ddisguise, "icon16/user.png", false, false, trans("equip_tooltip_disguise"))
 	end)
-else
+
+	hook.Add("Initialize", "TTTItemDisguiserInitStatus", function()
+		STATUS:RegisterStatus("item_disguiser_status", {
+			hud = Material("vgui/ttt/perks/hud_disguiser.png"),
+			type = "good"
+		})
+	end)
+end
+
+if SERVER then
 	local function SetDisguise(ply, cmd, args)
 		if not IsValid(ply) or not ply:IsActive() and ply:HasTeam(TEAM_TRAITOR) then return end
 
@@ -108,7 +89,19 @@ else
 
 		ply:SetNWBool("disguised", state)
 
-		LANG.Msg(ply, state and "disg_turned_on" or "disg_turned_off")
+		LANG.Msg(ply, state and "disg_turned_on" or "disg_turned_off", nil, MSG_MSTACK_ROLE)
 	end
 	concommand.Add("ttt_set_disguise", SetDisguise)
+
+	hook.Add("TTT2PlayerReady", "TTTItemDisguiserPlayerReady", function(ply)
+		ply:SetNWVarProxy("disguised", function(ent, name, old, new)
+			if not IsValid(ent) or not ent:IsPlayer() or name ~= "disguised" then return end
+
+			if new then
+				STATUS:AddStatus(ent, "item_disguiser_status")
+			else
+				STATUS:RemoveStatus(ent, "item_disguiser_status")
+			end
+		end)
+	end)
 end