+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 
 ### Changed
 
+- Moved the disguiser icon to the status system to be only displayed when the player is actually disguised
 - Reworked the addonchecker and added a command to execute the checker at a later point
 - Updated Italian translation (Thanks @ThePlatinumGhost)
 - Removed Is[ROLE] functions of all roles except default TTT ones